move diviews to views
